# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"製造業 AI 辨識系統","version":"0.1.0"},"paths":{"/api/general/describe":{"post":{"tags":["M9 開放式辨識"],"summary":"Describe","operationId":"describe_api_general_describe_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_describe_api_general_describe_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/docs/extract":{"post":{"tags":["M4 製造文件結構化"],"summary":"Extract","operationId":"extract_api_docs_extract_post","parameters":[{"name":"mode","in":"query","required":false,"schema":{"type":"string","default":"ocr","title":"Mode"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_extract_api_docs_extract_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/anomaly/inspect":{"post":{"tags":["M3 外觀瑕疵異常檢測"],"summary":"Inspect","operationId":"inspect_api_anomaly_inspect_post","parameters":[{"name":"category","in":"query","required":false,"schema":{"type":"string","default":"metal_nut","title":"Category"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_inspect_api_anomaly_inspect_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/defect/pcb":{"post":{"tags":["M6 監督式瑕疵偵測"],"summary":"Pcb","operationId":"pcb_api_defect_pcb_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_pcb_api_defect_pcb_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/codes/decode":{"post":{"tags":["M1 追溯碼辨識"],"summary":"Decode","operationId":"decode_api_codes_decode_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_decode_api_codes_decode_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/measure/count":{"post":{"tags":["M2 計數與尺寸量測"],"summary":"Count","operationId":"count_api_measure_count_post","parameters":[{"name":"expected_count","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Expected Count"}},{"name":"min_area_ratio","in":"query","required":false,"schema":{"type":"number","default":0.0005,"title":"Min Area Ratio"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_count_api_measure_count_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/measure/measure":{"post":{"tags":["M2 計數與尺寸量測"],"summary":"Measure","operationId":"measure_api_measure_measure_post","parameters":[{"name":"marker_size_mm","in":"query","required":false,"schema":{"type":"number","default":30.0,"title":"Marker Size Mm"}},{"name":"target_length_mm","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Target Length Mm"}},{"name":"target_width_mm","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Target Width Mm"}},{"name":"tolerance_mm","in":"query","required":false,"schema":{"type":"number","default":1.0,"title":"Tolerance Mm"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_measure_api_measure_measure_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/nameplate/read":{"post":{"tags":["M7 銘牌／儀表讀值"],"summary":"Nameplate","operationId":"nameplate_api_nameplate_read_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_nameplate_api_nameplate_read_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/nameplate/seven-segment":{"post":{"tags":["M7 銘牌／儀表讀值"],"summary":"Seven Segment","operationId":"seven_segment_api_nameplate_seven_segment_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_seven_segment_api_nameplate_seven_segment_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/nameplate/gauge":{"post":{"tags":["M7 銘牌／儀表讀值"],"summary":"Gauge","description":"角度慣例：0 度＝3 點鐘方向，順時針遞增（見 core/gauge.py 說明）。\ncenter_x/center_y 是正規化座標（0~1）；不給就自動找圓心。","operationId":"gauge_api_nameplate_gauge_post","parameters":[{"name":"min_value","in":"query","required":false,"schema":{"type":"number","default":0,"title":"Min Value"}},{"name":"max_value","in":"query","required":false,"schema":{"type":"number","default":100,"title":"Max Value"}},{"name":"min_angle","in":"query","required":false,"schema":{"type":"number","default":135,"title":"Min Angle"}},{"name":"max_angle","in":"query","required":false,"schema":{"type":"number","default":45,"title":"Max Angle"}},{"name":"center_x","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Center X"}},{"name":"center_y","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Center Y"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_gauge_api_nameplate_gauge_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/medical/packaging-check":{"post":{"tags":["M8 醫療相關辨識"],"summary":"Packaging Check","description":"比對 GS1 UDI 條碼跟印刷文字上的批號/效期是否一致（GMP 追溯用途）。","operationId":"packaging_check_api_medical_packaging_check_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_packaging_check_api_medical_packaging_check_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/medical/pneumonia-demo":{"post":{"tags":["M8 醫療相關辨識"],"summary":"Pneumonia Demo","description":"PneumoniaMNIST 教學展示分類。僅供技術展示，非醫療診斷用途。","operationId":"pneumonia_demo_api_medical_pneumonia_demo_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_pneumonia_demo_api_medical_pneumonia_demo_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/safety/intrusion":{"post":{"tags":["M5 工安：危險區域入侵 / PPE"],"summary":"Intrusion","description":"zone：JSON 字串，正規化座標（0~1）多邊形頂點，例如 '[[0.1,0.1],[0.5,0.1],[0.5,0.9]]'；不給就只做人員偵測。","operationId":"intrusion_api_safety_intrusion_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_intrusion_api_safety_intrusion_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/safety/ppe":{"post":{"tags":["M5 工安：危險區域入侵 / PPE"],"summary":"Ppe","description":"安全帽偵測；資料集只有 helmet/head 兩類，沒有反光背心。","operationId":"ppe_api_safety_ppe_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_ppe_api_safety_ppe_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/inspections":{"get":{"tags":["檢驗紀錄"],"summary":"List Inspections","operationId":"list_inspections_api_inspections_get","parameters":[{"name":"module","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Module"}},{"name":"verdict","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Verdict"}},{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date From"}},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date To"}},{"name":"work_order","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Work Order"}},{"name":"part_no","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Part No"}},{"name":"lot_no","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Lot No"}},{"name":"station","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Station"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":200,"title":"Limit"}},{"name":"since_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Since Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/inspections/reviews":{"get":{"tags":["檢驗紀錄"],"summary":"List Reviews","description":"給 ERP 抓「事後被改判」的紀錄：since 是 ISO 時間字串（例如 2026-09-24T00:00:00）。","operationId":"list_reviews_api_inspections_reviews_get","parameters":[{"name":"since","in":"query","required":true,"schema":{"type":"string","title":"Since"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":200,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/inspections/export.csv":{"get":{"tags":["檢驗紀錄"],"summary":"Export Csv","operationId":"export_csv_api_inspections_export_csv_get","parameters":[{"name":"module","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Module"}},{"name":"verdict","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Verdict"}},{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date From"}},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date To"}},{"name":"work_order","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Work Order"}},{"name":"part_no","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Part No"}},{"name":"lot_no","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Lot No"}},{"name":"station","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Station"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10000,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/inspections/stats":{"get":{"tags":["檢驗紀錄"],"summary":"Stats","operationId":"stats_api_inspections_stats_get","parameters":[{"name":"group_by","in":"query","required":false,"schema":{"type":"string","default":"module","title":"Group By"}},{"name":"date_from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date From"}},{"name":"date_to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Date To"}},{"name":"module","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Module"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/inspections/{inspection_id}/review":{"patch":{"tags":["檢驗紀錄"],"summary":"Review Inspection","operationId":"review_inspection_api_inspections__inspection_id__review_patch","parameters":[{"name":"inspection_id","in":"path","required":true,"schema":{"type":"integer","title":"Inspection Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/inspections/{inspection_id}":{"get":{"tags":["檢驗紀錄"],"summary":"Get One","operationId":"get_one_api_inspections__inspection_id__get","parameters":[{"name":"inspection_id","in":"path","required":true,"schema":{"type":"integer","title":"Inspection Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/inspections/{inspection_id}/image":{"get":{"tags":["檢驗紀錄"],"summary":"Get Image","operationId":"get_image_api_inspections__inspection_id__image_get","parameters":[{"name":"inspection_id","in":"path","required":true,"schema":{"type":"integer","title":"Inspection Id"}},{"name":"kind","in":"query","required":false,"schema":{"type":"string","default":"raw","title":"Kind"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_count_api_measure_count_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_count_api_measure_count_post"},"Body_decode_api_codes_decode_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_decode_api_codes_decode_post"},"Body_describe_api_general_describe_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_describe_api_general_describe_post"},"Body_extract_api_docs_extract_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_extract_api_docs_extract_post"},"Body_gauge_api_nameplate_gauge_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_gauge_api_nameplate_gauge_post"},"Body_inspect_api_anomaly_inspect_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_inspect_api_anomaly_inspect_post"},"Body_intrusion_api_safety_intrusion_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"},"zone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Zone"}},"type":"object","required":["file"],"title":"Body_intrusion_api_safety_intrusion_post"},"Body_measure_api_measure_measure_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_measure_api_measure_measure_post"},"Body_nameplate_api_nameplate_read_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_nameplate_api_nameplate_read_post"},"Body_packaging_check_api_medical_packaging_check_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_packaging_check_api_medical_packaging_check_post"},"Body_pcb_api_defect_pcb_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_pcb_api_defect_pcb_post"},"Body_pneumonia_demo_api_medical_pneumonia_demo_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_pneumonia_demo_api_medical_pneumonia_demo_post"},"Body_ppe_api_safety_ppe_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_ppe_api_safety_ppe_post"},"Body_seven_segment_api_nameplate_seven_segment_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_seven_segment_api_nameplate_seven_segment_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InspectionResult":{"properties":{"module":{"type":"string","title":"Module"},"verdict":{"type":"string","enum":["OK","NG","INFO"],"title":"Verdict"},"items":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Items"},"annotated_image":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Annotated Image"},"engine":{"type":"string","title":"Engine"},"elapsed_ms":{"type":"integer","title":"Elapsed Ms"},"inspection_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Inspection Id"}},"type":"object","required":["module","verdict","items","engine","elapsed_ms"],"title":"InspectionResult"},"ReviewRequest":{"properties":{"review_verdict":{"type":"string","title":"Review Verdict","description":"OK 或 NG"},"reviewer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reviewer"},"review_note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Review Note"}},"type":"object","required":["review_verdict"],"title":"ReviewRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}